requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
Pillow>=10.0.0,<11.0.0
